- 詰み関係の処理を分離。     source/extra/mate/ を     source/mate/ に移動   - Positionクラスから詰め将棋関係の関数をnamespace Mateに移動。   - mate/mate.cpp追加。
