+++ source/Makefile
@@ -233,12 +233,13 @@ SOURCES  = \
 	extra/bitop.cpp                                                            \
 	extra/long_effect.cpp                                                      \
 	extra/mate/mate1ply_with_effect.cpp                                        \
-	extra/mate/mate1ply_without_effect.cpp                                     \
-	extra/mate/mate_n_ply.cpp                                                  \
 	extra/test_cmd.cpp                                                         \
 	extra/sfen_packer.cpp                                                      \
 	extra/kif_converter/kif_convert_tools.cpp                                  \
 	extra/super_sort.cpp                                                       \
+	mate/mate.cpp                                                              \
+	mate/mate1ply_without_effect.cpp                                           \
+	mate/mate_n_ply.cpp                                                        \
 	eval/evaluate_bona_piece.cpp                                               \
 	eval/evaluate.cpp                                                          \
 	eval/evaluate_io.cpp                                                       \
