+++ source/YaneuraOu.vcxproj
@@ -551,11 +551,11 @@
     <ClInclude Include="extra\kif_converter\kif_convert_tools.h" />
     <ClInclude Include="extra\long_effect.h" />
     <ClInclude Include="extra\macros.h" />
-    <ClInclude Include="extra\mate\mate1ply.h" />
     <ClInclude Include="learn\half_float.h" />
     <ClInclude Include="learn\learn.h" />
     <ClInclude Include="learn\learning_tools.h" />
     <ClInclude Include="learn\multi_think.h" />
+    <ClInclude Include="mate\mate.h" />
     <ClInclude Include="misc.h" />
     <ClInclude Include="movepick.h" />
     <ClInclude Include="search.h" />
@@ -608,9 +608,6 @@
     <ClCompile Include="extra\bitop.cpp" />
     <ClCompile Include="extra\kif_converter\kif_convert_tools.cpp" />
     <ClCompile Include="extra\long_effect.cpp" />
-    <ClCompile Include="extra\mate\mate1ply_without_effect.cpp" />
-    <ClCompile Include="extra\mate\mate1ply_with_effect.cpp" />
-    <ClCompile Include="extra\mate\mate_n_ply.cpp" />
     <ClCompile Include="extra\sfen_packer.cpp" />
     <ClCompile Include="extra\super_sort.cpp" />
     <ClCompile Include="extra\test_cmd.cpp" />
@@ -618,6 +615,10 @@
     <ClCompile Include="learn\learning_tools.cpp" />
     <ClCompile Include="learn\multi_think.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="mate\mate.cpp" />
+    <ClCompile Include="mate\mate1ply_without_effect.cpp" />
+    <ClCompile Include="mate\mate1ply_with_effect.cpp" />
+    <ClCompile Include="mate\mate_n_ply.cpp" />
     <ClCompile Include="misc.cpp" />
     <ClCompile Include="movegen.cpp" />
     <ClCompile Include="movepick.cpp" />
