+++ source/extra/all.h
@@ -20,6 +20,7 @@
 #include "../book/book.h"
 #include "../learn/learn.h"
 #include "kif_converter/kif_convert_tools.h"
+#include "../mate/mate.h"
 
 // これもおまけしておく。
 using namespace std;
